Moved Button to its own file, Added Connect4 -> MainMenu transition & button; Refactored reset button to use new Button
