+++ src/connect4/core.rs
@@ -9,6 +9,7 @@ use ggez::mint::Point2;
 use ggez::input::mouse;
 use ggez::input::mouse::MouseButton;
 use connect4::ai::AI;
+use connect4::button::{BUTTON_PADDING, BUTTON_SPACING, Button};
 
 /// Constant definition for the connect4 board size: 6x7 cells, row x column
 pub const BOARD_SIZE: (i32, i32) = (6, 7);
@@ -51,7 +52,6 @@ pub enum MyColor {
     White,
     Blue,
     Red,
-    Green
     Green,
     Brown,
     Black,
@@ -63,7 +63,6 @@ impl MyColor {
             MyColor::White => graphics::WHITE,
             MyColor::Blue => graphics::Color::from_rgba(0,0,255,255),
             MyColor::Red => graphics::Color::from_rgba(255,0,0,255),
-            MyColor::Green => graphics::Color::from_rgba(0,255,0,255)
             MyColor::Green => graphics::Color::from_rgba(0,255,0,255),
             MyColor::Brown => graphics::Color::from_rgba(205,133,63,255),
             MyColor::Black => graphics::BLACK,
@@ -485,42 +484,6 @@ impl TurnIndicator {
     }
 }
 
-// Button struct from original main
-pub struct Button {
-    text: graphics::Text,
-    outline: graphics::Rect,
-    highlighted: bool,
-}
-
-impl Button {
-    pub fn new(text: graphics::Text, dim: graphics::Rect) -> Button {
-        Button { text: text, outline: dim, highlighted: false}
-    }
-
-    pub fn draw(&self, ctx: &mut Context) -> GameResult<()> {
-        let textbox = graphics::Mesh::new_rectangle(
-            ctx, 
-            graphics::DrawMode::fill(),             
-            self.outline,
-            graphics::Color::from_rgba(133,0,0,255),
-        )?;
-        graphics::draw(ctx, &textbox, (Point2 {x: 0.0, y: 0.0},))?;
-        graphics::draw(ctx, &self.text, (Point2 {x: RESET_BUTTON_OFFSET.0 as f32, y: RESET_BUTTON_OFFSET.1 as f32},))?;
-        Ok(())
-    }
-
-    pub fn is_button_under_mouse(&mut self, ctx: &mut Context) -> bool {
-        let mouse_loc = mouse::position(ctx);
-        if self.outline.contains(mouse_loc)  {
-            self.highlighted = true;
-        } else {
-            self.highlighted = false;
-        }
-        self.highlighted
-    }
-
-}
-
 pub struct GameState {
     frames: usize,
     ai_players: Vec<AI>,
@@ -531,16 +494,25 @@ pub struct GameState {
     mouse_disabled: bool,
     gameover: bool,
     pub reset_button: Button,
-
+    pub main_menu_button: Button,
 }
 
 //Implementation based on structure in example from GGEZ repo (see https://github.com/ggez/ggez/blob/master/examples/02_hello_world.rs)
 impl GameState {
     pub fn new(ctx: &mut Context, players: i32) -> GameState {
         let board_pos = BOARD_POS_OFFSET;
-        let text = graphics::Text::new("Reset");
-        let text_width = text.width(ctx) as f32;
-        let text_height = text.height(ctx) as f32;
+        let main_menu_btn_text = graphics::Text::new(("Main Menu", graphics::Font::default(), 16f32,));
+        let main_menu_text_width = main_menu_btn_text.width(ctx) as f32;
+        let main_menu_text_height = main_menu_btn_text.height(ctx) as f32;
+        let main_menu_btn_outline = graphics::Rect::new(RESET_BUTTON_OFFSET.0 as f32, RESET_BUTTON_OFFSET.1 as f32 + main_menu_text_height, main_menu_text_width, main_menu_text_height);
+        let mut main_menu_btn = Button::new(main_menu_btn_text, main_menu_btn_outline);
+
+        let reset_text = graphics::Text::new(("Reset", graphics::Font::default(), 16f32,));
+        let reset_outline = graphics::Rect::new(RESET_BUTTON_OFFSET.0 as f32, RESET_BUTTON_OFFSET.1 as f32 + main_menu_text_height*3.0, main_menu_text_width, main_menu_text_height);
+        let mut reset_btn = Button::new(reset_text, reset_outline);
+
+        reset_btn.set_colors(MyColor::Brown, MyColor::Red);
+        main_menu_btn.set_colors(MyColor::Brown, MyColor::Green);
         let mut bots = Vec::<AI>::new();
         for i in 0..players {
             bots.push(AI::new(2-i, 3));
@@ -554,7 +526,8 @@ impl GameState {
             highlighted_column: -1,
             mouse_disabled: false,
             gameover: false,
-            reset_button: Button::new(text, graphics::Rect::new(RESET_BUTTON_OFFSET.0 as f32, RESET_BUTTON_OFFSET.1 as f32, text_width, text_height)),
+            reset_button: reset_btn,
+            main_menu_button: main_menu_btn,
         }
     }
 
@@ -606,6 +579,7 @@ impl GameState {
 
         //Draw reset button
         self.reset_button.draw(ctx)?;
+        self.main_menu_button.draw(ctx)?;
         graphics::present(ctx)?;
         ggez::timer::yield_now();
         Ok(())
@@ -620,16 +594,20 @@ impl GameState {
                 println!("Mouse moved to col {}", self.highlighted_column);
             }
         }
+        self.reset_button.is_button_under_mouse(_ctx);
+        self.main_menu_button.is_button_under_mouse(_ctx);
     }
 
     pub fn mouse_button_down_event(&mut self, _ctx: &mut Context, _button: MouseButton, _x: f32, _y: f32) {
         if !self.mouse_disabled {
             self.highlighted_column = self.board.get_highlighted_column(mouse::position(_ctx));
         }
+        self.reset_button.is_button_under_mouse(_ctx);
+        self.main_menu_button.is_button_under_mouse(_ctx);
     }
 
     //Todo: If mouse_motion_event is enabled, this will always drop the token (i.e. not click away to undo move). Undetermined if this is desired or not
-    pub fn mouse_button_up_event(&mut self, _ctx: &mut Context, _button: MouseButton, _x: f32, _y: f32) {
+    pub fn mouse_button_up_event(&mut self, _ctx: &mut Context, _button: MouseButton, _x: f32, _y: f32) -> bool {
         if !self.mouse_disabled {
             let was_highlighted = self.highlighted_column;
             self.highlighted_column = self.board.get_highlighted_column(mouse::position(_ctx));
@@ -662,6 +640,13 @@ impl GameState {
             self.turnIndicator.change_team(1);
             self.gameover = false;
             self.mouse_disabled = false;
+        }
+
+        if self.main_menu_button.is_button_under_mouse(_ctx) {
+            println!("Main Menu Button pressed; Main Menu should pop up");
+            true
+        } else {
+            false
         }
     }
 }
