+++ src/connect4/mod.rs
@@ -5,3 +5,4 @@
 
 pub mod core;
 pub mod ai;
+pub mod button;
